Some rights problem fixed.

+++ admin/genfiles/gen_maildrop_userdb.php
@@ -84,7 +84,13 @@ function mail_account_generate_maildrop(){
 
 			system("maildirmake -q  $quota_maildrop $boxpath/Maildir");
 			if($quota_maildrop==0){
-				system("rm $boxpath/Maildir/maildirsize");
+				if(file_exists("$boxpath/Maildir/maildirsize")){
+					system("rm $boxpath/Maildir/maildirsize");
+				}
+			}else{
+				if($panel_type == "cronjob"){
+					chown("$boxpath/Maildir/maildirsize",$conf_dtc_system_username);
+				}
 			}
 
 		}
